Reverted change to how admin profile fields are pulled since they didn't need a correction
With the changes made elsewhere the $volunteer->meta property is already
using get_user_option() to pull the phone number and notes, so no need
to adjust the lines in class-admin-php. The phone number also has
additional formatting that is done in the property, so pulling it
directly would remove that formatting.

The changes made to class-volunteer.php are just for consistency.
Nothing important there.